backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.2

# Vector Database
chromadb==0.4.18

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
langchain-text-splitters==0.0.1

# Streamlit for frontend
streamlit==1.28.2

# Additional utilities
requests==2.31.0
